test(palette-service): correctly mock palette service for tests
