+++ src/app/shared/data-access/confetti.service.spec.ts
@@ -1,6 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { ConfettiService } from './confetti.service';
-import { PaletteService, PaletteServiceMock } from './palette.service';
+import { PaletteService } from './palette.service';
+import { PaletteServiceMock } from './palette.service-mock';
 
 describe('ConfettiService', () => {
   let service: ConfettiService;
